[Feat] #83 - rewrite __init__.py
