+++ chzzkpy/__init__.py
@@ -28,3 +28,35 @@ from .client import Client, UserClient
 from .error import *
 from .flags import UserPermission
 from .message import Donation, Profile, Message
+
+
+try:
+    from .unoffical import *
+except ModuleNotFoundError:
+    pass
+
+
+__title__ = "chzzkpy"
+__author__ = "gunyu1019"
+__license__ = "MIT"
+__copyright__ = "Copyright 2024-present gunyu1019"
+__version__ = "2.1.0-beta1"  # version_info.to_string()
+
+
+class VersionInfo(NamedTuple):
+    major: int
+    minor: int
+    micro: int
+    release_level: Optional[Literal["alpha", "beta", "candidate", "final"]]
+    serial: int
+
+    def to_string(self) -> str:
+        _version_info = f"{self.major}.{self.minor}.{self.micro}"
+        if self.release_level is not None:
+            _version_info += "-{}".format(self.release_level) + str(self.serial)
+        return _version_info
+
+
+version_info: VersionInfo = VersionInfo(
+    major=2, minor=1, micro=0, release_level="beta", serial=1
+)
