+++ chzzkpy/unoffical/__init__.py
@@ -34,34 +34,3 @@ try:
     from .chat import *
 except ModuleNotFoundError:
     pass
-
-try:
-    from .offical import *
-except ModuleNotFoundError:
-    pass
-
-
-__title__ = "chzzkpy"
-__author__ = "gunyu1019"
-__license__ = "MIT"
-__copyright__ = "Copyright 2024-present gunyu1019"
-__version__ = "2.0.6"  # version_info.to_string()
-
-
-class VersionInfo(NamedTuple):
-    major: int
-    minor: int
-    micro: int
-    release_level: Optional[Literal["alpha", "beta", "candidate", "final"]]
-    serial: int
-
-    def to_string(self) -> str:
-        _version_info = f"{self.major}.{self.minor}.{self.micro}"
-        if self.release_level is not None:
-            _version_info += "-{}".format(self.release_level) + str(self.serial)
-        return _version_info
-
-
-version_info: VersionInfo = VersionInfo(
-    major=2, minor=0, micro=6, release_level="", serial=0
-)
